reorganize code and files
