--- DMR/PaddleRec/models/rank/dmr/alimama_reader.py
@@ -1,33 +1,0 @@
-#   Copyright (c) 2020 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from __future__ import print_function
-import numpy as np
-
-from paddle.io import IterableDataset
-
-
-class RecDataset(IterableDataset):
-    def __init__(self, file_list, config):
-        super(RecDataset, self).__init__()
-        self.file_list = file_list
-
-    def __iter__(self):
-        for file in self.file_list:
-            with open(file, "r") as rf:
-                for l in rf:
-                    l = l.strip().split(",")
-                    l = ['0' if i=='' or i.upper()=='NULL' else i for i in l]  # handle missing values
-                    output_list = np.array(l, dtype='float32')
-                    yield output_list
