+++ DMR/requirements.txt
@@ -1,0 +1,1 @@
+paddlepaddle==2.1.0
